requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
python-pptx
pymupdf
numpy
langchain
langchain-community
sentence-transformers
langchain_openai
langchain-google-genai
google-genai